shopify_Janus_backend/requirements-test.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
gradio-client==0.2.7
pytest==6.2.5
pytest-cov==2.12.1 